+++ server/commands.go
@@ -18,14 +18,15 @@ import (
 type command func(c *Connection, s *Server, cmd string)
 
 var cmdSet = map[string]command{
-	"END":   end,
-	"PUSH":  push,
-	"FETCH": fetch,
-	"ACK":   ack,
-	"FAIL":  fail,
-	"BEAT":  heartbeat,
-	"INFO":  info,
-	"FLUSH": flush,
+	"END":    end,
+	"PUSH":   push,
+	"FETCH":  fetch,
+	"ACK":    ack,
+	"FAIL":   fail,
+	"BEAT":   heartbeat,
+	"INFO":   info,
+	"FLUSH":  flush,
+	"MUTATE": mutate,
 }
 
 func flush(c *Connection, s *Server, cmd string) {