+++ storage/types.go
@@ -81,12 +81,15 @@ type SortedSet interface {
 	Page(start int, count int, fn func(index int, e SortedEntry) error) (int, error)
 	Each(fn func(idx int, e SortedEntry) error) error
 
+	Find(match string, fn func(idx int, e SortedEntry) error) error
+
 	// bool is whether or not the element was actually removed from the sset.
 	// the scheduler and other things can be operating on the sset concurrently
 	// so we need to be careful about the data changing under us.
 	Remove(key []byte) (bool, error)
 	RemoveElement(timestamp string, jid string) (bool, error)
 	RemoveBefore(timestamp string) ([][]byte, error)
+	RemoveEntry(ent SortedEntry) error
 
 	// Move the given key from this SortedSet to the given
 	// SortedSet atomically.  The given func may mutate the payload and